+++ common/src/main/java/com/st0x0ef/stellaris/common/data_components/RocketComponent.java
@@ -12,18 +12,18 @@ import net.minecraft.resources.ResourceLocation;
 
 import java.io.Serializable;
 
-public record RocketComponent(String skin, RocketModel model, String fuelType, int fuel, int tankCapacity) implements Serializable {
+public record RocketComponent(String skin, RocketModel model, String fuelType, int fuel, ResourceLocation fuelTexture, int tankCapacity) implements Serializable {
 
     public static final Codec<RocketComponent> CODEC = RecordCodecBuilder.create(instance -> instance.group(
             Codec.STRING.fieldOf("skin").forGetter(RocketComponent::skin),
             RocketModel.CODEC.fieldOf("model").forGetter(RocketComponent::model),
             Codec.STRING.fieldOf("fuel_type").forGetter(RocketComponent::fuelType),
             Codec.INT.fieldOf("fuel").forGetter(RocketComponent::fuel),
+            ResourceLocation.CODEC.fieldOf("fuel_texture").forGetter(RocketComponent::fuelTexture),
             Codec.INT.fieldOf("fuel_capacity").forGetter(RocketComponent::tankCapacity)
     ).apply(instance, RocketComponent::new));
 
     public static final StreamCodec<ByteBuf, RocketComponent> STREAM_CODEC;
-
 
     public int getFuel() {
         return fuel;
@@ -46,7 +46,7 @@ public record RocketComponent(String skin, RocketModel model, String fuelType, i
     }
 
     public MotorUpgrade getMotorUpgrade() {
-        return new MotorUpgrade(FuelType.Type.fromString(fuelType));
+        return new MotorUpgrade(FuelType.Type.fromString(fuelType), fuelTexture);
     }
 
     public TankUpgrade getTankUpgrade() {
@@ -59,7 +59,7 @@ public record RocketComponent(String skin, RocketModel model, String fuelType, i
     }
 
     public static RocketComponent fromNetwork(RegistryFriendlyByteBuf buffer) {
-        return new RocketComponent(buffer.readUtf(), RocketModel.fromString(buffer.readUtf()), buffer.readUtf(), buffer.readInt(), buffer.readInt());
+        return new RocketComponent(buffer.readUtf(), RocketModel.fromString(buffer.readUtf()), buffer.readUtf(), buffer.readInt(), buffer.readResourceLocation(), buffer.readInt());
     }
 
     public RegistryFriendlyByteBuf toNetwork(RegistryFriendlyByteBuf buffer) {
@@ -67,12 +67,13 @@ public record RocketComponent(String skin, RocketModel model, String fuelType, i
         buffer.writeUtf(this.model().getSerializedName());
         buffer.writeUtf(this.fuelType);
         buffer.writeInt(this.fuel);
+        buffer.writeResourceLocation(this.fuelTexture);
         buffer.writeInt(this.tankCapacity);
         return buffer;
     }
 
 
     static {
-        STREAM_CODEC = StreamCodec.composite(ByteBufCodecs.STRING_UTF8, RocketComponent::skin, ByteBufCodecs.fromCodec(RocketModel.CODEC), RocketComponent::model, ByteBufCodecs.STRING_UTF8, RocketComponent::fuelType, ByteBufCodecs.INT, RocketComponent::fuel, ByteBufCodecs.INT, RocketComponent::tankCapacity, RocketComponent::new);
+        STREAM_CODEC = StreamCodec.composite(ByteBufCodecs.STRING_UTF8, RocketComponent::skin, ByteBufCodecs.fromCodec(RocketModel.CODEC), RocketComponent::model, ByteBufCodecs.STRING_UTF8, RocketComponent::fuelType, ByteBufCodecs.INT, RocketComponent::fuel, ResourceLocation.STREAM_CODEC, RocketComponent::fuelTexture, ByteBufCodecs.INT, RocketComponent::tankCapacity, RocketComponent::new);
     }
 }